+++ src/kernel/none/level1.h
@@ -814,7 +814,7 @@ INLINE long
 expi(GEN x)
 {
   const long lx=lgefint(x);
-  return lx==2? -HIGHEXPOBIT: bit_accuracy(lx)-bfffo(x[2])-1;
+  return lx==2? -(long)HIGHEXPOBIT: bit_accuracy(lx)-bfffo(x[2])-1;
 }
 
 #endif
